Fixed import/no-cycle issues
